added pagination to watches

+++ app/helpers/users/lists_helper.rb
@@ -1,7 +1,7 @@
 module Users::ListsHelper
   def link_to_watch(list)
     if current_user.watch?(list)
-      link_to "Unwatch", watch_path(current_user.watch(list)), :method => :delete, :title => "Unwatch"
+      link_to "Unwatch", watch_path(current_user.watch(list)), :method => :delete, :confirm => "Are you sure?", :title => "Unwatch"
     else
       link_to "Watch", watches_path(:watch => { :list_id => list.id }), :method => :post, :title => "Watch"
     end
